+++ app/javascript/src/components/Reports/AccountsAgingReport/Container/Table/TableHeader.tsx
