+++ app/javascript/src/components/Reports/AccountsAgingReport/Container/Table/TableTotal.tsx
